src/uvplayer_monitor/requirements.txt
pyautogui
pillow
psutil
pywin32
opencv-python
numpy
